perf(bundle): defer dev-only key-collision warnings to a dynamic-imported module
The __DEV__-only shared-key collision diagnostics (warnOnSchemaFingerprint-
Mismatch, warnOnPersistDivergence, and two helpers) shipped ~0.47 kB gz of
dead code into production. esbuild drops an inline `if (__DEV__)` block but
keeps a top-level function called only from one, because tree-shaking runs
before the define-fold, so the orphaned function survives.

Move the cluster to a dev-only module (dev-key-collision-warnings.ts) loaded
via a `__DEV__`-gated dynamic import, mirroring the devtools pattern. A prod
build folds the gate, DCEs the import() call, and orphans the chunk, so none
of it ships. Eager drops 44.38 -> 43.91 kB gz; check-eager-size's budget
tightens to lock it in.

dev-dce gains an F3 case asserting the module is reachable in dev and
unreachable in prod. schema-fingerprint-warning warms the module in beforeAll
so its first dynamic import resolves from cache (one extra hop otherwise
lands after the test's timer drain).
